+++ dashboard.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
     <title>Dashboard - Brand</title>
     <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
+    <link rel="stylesheet" href="assets/css/styles.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i">
     <link rel="stylesheet" href="assets/fonts/fontawesome-all.min.css">
     <link rel="stylesheet" href="assets/fonts/font-awesome.min.css">
@@ -14,10 +15,10 @@
     <link rel="stylesheet" href="assets/css/iframe.css">
     <link rel="stylesheet" href="assets/css/Login-Form-Clean.css">
     <link rel="stylesheet" href="assets/css/Select2-JS-Plugin.css">
-    <link rel="stylesheet" href="assets/css/styles.css">
 </head>
 
 <body id="page-top">
+    <div id="loading"></div>
     <div id="wrapper">
         <nav class="navbar navbar-dark align-items-start sidebar sidebar-dark accordion bg-gradient-primary p-0" style="background: #ff5200;">
             <div class="container-fluid d-flex flex-column p-0"><a class="navbar-brand d-flex justify-content-center align-items-center sidebar-brand m-0" href="#">
@@ -63,7 +64,8 @@
                 </nav>
                 <div class="container-fluid">
                     <div class="d-sm-flex justify-content-between align-items-center mb-4">
-                        <h3 class="text-dark mb-0">Dashboard</h3><button class="btn btn-primary" id="add-drawings" type="button" data-target="#drawings-modal" data-toggle="modal">Add Drawings</button>
+                        <h3 class="text-dark mb-0">Dashboard</h3>
+                        <div><button class="btn btn-primary" id="add-drawings" type="button" data-target="#drawings-modal" data-toggle="modal">Add Drawings</button><button class="btn btn-primary" id="updates-button" type="button" style="margin-left: 5px;">Check for Updates&nbsp;</button></div>
                     </div>
                     <div class="row">
                         <div class="col-lg-12 col-xl-12 mb-4">
@@ -134,7 +136,7 @@
                         <div class="progress-bar progress-bar-striped progress-bar-animated" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100" id="progress-bar" style="width: 25%;">25%</div>
                     </div>
                 </div>
-                <div class="modal-footer"><button class="btn btn-light" type="button" data-dismiss="modal">Close</button><button class="btn btn-primary" id="cancel-button" type="button">Cancel</button></div>
+                <div class="modal-footer"><button class="btn btn-primary" type="button" data-dismiss="modal">Close</button></div>
             </div>
         </div>
     </div>
@@ -151,7 +153,6 @@
             </div>
         </div>
     </div>
-    <div id="loading"></div>
     <script>if (typeof module === 'object') { window.module = module; module = undefined; }</script>
     <script>var exports = { "__esModule": true };</script>
     <script src="assets/js/jquery.min.js"></script>
