+++ login.html
@@ -14,7 +14,6 @@
     <link rel="stylesheet" href="assets/css/iframe.css">
     <link rel="stylesheet" href="assets/css/Login-Form-Clean.css">
     <link rel="stylesheet" href="assets/css/Select2-JS-Plugin.css">
-    <link rel="stylesheet" href="assets/css/styles.css">
 </head>
 
 <body class="bg-gradient-primary" style="background: linear-gradient(#ff5200, white), #ff5200;margin-top: 128px;">
