+++ profile.html
@@ -14,7 +14,6 @@
     <link rel="stylesheet" href="assets/css/iframe.css">
     <link rel="stylesheet" href="assets/css/Login-Form-Clean.css">
     <link rel="stylesheet" href="assets/css/Select2-JS-Plugin.css">
-    <link rel="stylesheet" href="assets/css/styles.css">
 </head>
 
 <body id="page-top">
